requirements.txt
asgiref==3.3.4
bandit==1.7.0
Django==3.2.4
django-cors-headers==3.7.0
django-debug-toolbar==3.2.1
django-environ==0.4.5
django-filter==2.4.0
django-jazzmin==2.4.7
django-redis==5.0.0
djangorestframework==3.12.4
gitdb==4.0.7
GitPython==3.1.18
pbr==5.6.0
psycopg2==2.9.1
psycopg2-binary==2.9.1
pytz==2021.1
PyYAML==5.4.1
redis==3.5.3
six==1.16.0
smmap==4.0.0
sqlparse==0.4.1
stevedore==3.3.0
